[IMP] *: test perfs
Focus on improving the performances for the test suite. Some actions are
not often done in production (e.g. creating a company) but very often in
tests.
This is important to get a faster test build time.

+++ addons/account/models/account_bank_statement_line.py
@@ -456,6 +456,8 @@ class AccountBankStatementLine(models.Model):
         """ Undo the reconciliation made on the statement line and reset their journal items
         to their original states.
         """
+        if not self:
+            return
         self.line_ids.remove_move_reconcile()
         self.payment_ids.unlink()
